+++ src/restfw/tests/test_viewderivers.py
@@ -3,7 +3,8 @@
 :Authors: cykooz
 :Date: 27.12.2019
 """
-from pyramid.httpexceptions import HTTPMethodNotAllowed
+import pytest
+from pyramid.httpexceptions import HTTPMethodNotAllowed, HTTPNotModified, HTTPPreconditionFailed
 from pyramid.security import ALL_PERMISSIONS, Allow, Everyone
 from pyramid.view import view_config
 
@@ -11,6 +12,7 @@ from .. import schemas
 from ..errors import ResultValidationError
 from ..interfaces import MethodOptions
 from ..resources import Resource
+from ..utils import ETag
 
 
 class DummySchema(schemas.MappingSchema):
@@ -24,14 +26,16 @@ class DummyEditSchema(schemas.MappingSchema):
 
 
 class DummyResource(Resource):
-
     __acl__ = [
         (Allow, Everyone, ALL_PERMISSIONS)
     ]
 
     def __init__(self, model):
-        super(DummyResource, self).__init__()
         self.model = model
+        self.etag = None
+
+    def get_etag(self):
+        return self.etag
 
     options_for_get = MethodOptions(None, DummySchema, permission='get')
 
@@ -96,3 +100,144 @@ def test_check_result_schema(web_app, pyramid_request, app_config):
     custom_view_url = resource_url + 'custom_view'
     res = web_app.get(custom_view_url)
     assert res.json_body == {'custom': 'view', 'result': 789}
+
+
+@pytest.mark.parametrize(
+    ['etag', 'if_match', 'if_none_match', 'status_code'],
+    [
+        (None, None, None, 200),
+        (None, '*', None, 200),
+        (None, '"etag"', None, 412),
+        (None, None, '"etag"', 200),
+        (None, None, '*', 200),
+
+        (ETag('etag'), None, None, 200),
+        (ETag('etag'), '*', None, 200),
+        (ETag('etag'), '"etag"', None, 200),
+        (ETag('etag'), '"other", "etag"', None, 200),
+        (ETag('etag'), '"other"', None, 412),
+        (ETag('etag'), 'W/"etag"', None, 412),
+
+        (ETag('etag', False), None, None, 200),
+        (ETag('etag', False), '*', None, 200),
+        (ETag('etag', False), '"etag"', None, 412),
+        (ETag('etag', False), '"other", "etag"', None, 412),
+        (ETag('etag', False), '"other"', None, 412),
+        (ETag('etag', False), 'W/"etag"', None, 412),
+
+        (ETag('etag'), None, '"etag"', 304),
+        (ETag('etag'), None, '"other", "etag"', 304),
+        (ETag('etag'), None, 'W/"etag"', 304),
+        (ETag('etag'), None, '"other", W/"etag"', 304),
+        (ETag('etag'), None, '*', 304),
+        (ETag('etag'), None, '"other"', 200),
+
+        (ETag('etag', False), None, '"etag"', 304),
+        (ETag('etag', False), None, '"other", "etag"', 304),
+        (ETag('etag', False), None, 'W/"etag"', 304),
+        (ETag('etag', False), None, '"other", W/"etag"', 304),
+        (ETag('etag', False), None, '*', 304),
+        (ETag('etag', False), None, '"other"', 200),
+
+        (ETag('etag'), '"etag"', '"etag"', 304),
+        (ETag('etag'), '"etag"', 'W/"etag"', 304),
+        (ETag('etag'), '"etag"', '"other"', 200),
+        (ETag('etag'), 'W/"etag"', '"etag"', 412),
+        (ETag('etag'), 'W/"etag"', '"other"', 412),
+    ]
+)
+def test_process_conditional_get_head_requests(web_app, pyramid_request, app_config,
+                                               etag, if_match, if_none_match, status_code):
+    app_config.scan('restfw.tests.test_viewderivers')
+    app_config.commit()
+    root = pyramid_request.root
+    root['resource'] = resource = DummyResource({'foo': 'Hello', 'bar': 123})
+    resource.etag = etag
+    resource_url = pyramid_request.resource_url(root['resource'])
+
+    headers = {}
+    if if_match:
+        headers['If-Match'] = if_match
+    if if_none_match:
+        headers['If-None-Match'] = if_none_match
+
+    kwargs = {'headers': headers}
+    if status_code == 304:
+        kwargs['exception'] = HTTPNotModified()
+    elif status_code == 412:
+        kwargs['exception'] = HTTPPreconditionFailed({'etag': None if etag is None else etag.serialize()})
+    else:
+        kwargs['status'] = status_code
+
+    web_app.get(resource_url, **kwargs)
+
+
+@pytest.mark.parametrize(
+    ['etag', 'if_match', 'if_none_match', 'status_code'],
+    [
+        (None, None, None, 200),
+        (None, '*', None, 200),
+        (None, '"etag"', None, 412),
+        (None, None, '"etag"', 200),
+        (None, None, '*', 200),
+
+        (ETag('etag'), None, None, 200),
+        (ETag('etag'), '*', None, 200),
+        (ETag('etag'), '"etag"', None, 200),
+        (ETag('etag'), '"other", "etag"', None, 200),
+        (ETag('etag'), '"other"', None, 412),
+        (ETag('etag'), 'W/"etag"', None, 412),
+
+        (ETag('etag', False), None, None, 200),
+        (ETag('etag', False), '*', None, 200),
+        (ETag('etag', False), '"etag"', None, 412),
+        (ETag('etag', False), '"other", "etag"', None, 412),
+        (ETag('etag', False), '"other"', None, 412),
+        (ETag('etag', False), 'W/"etag"', None, 412),
+
+        (ETag('etag'), None, '"etag"', 412),
+        (ETag('etag'), None, '"other", "etag"', 412),
+        (ETag('etag'), None, 'W/"etag"', 412),
+        (ETag('etag'), None, '"other", W/"etag"', 412),
+        (ETag('etag'), None, '*', 412),
+        (ETag('etag'), None, '"other"', 200),
+
+        (ETag('etag', False), None, '"etag"', 412),
+        (ETag('etag', False), None, '"other", "etag"', 412),
+        (ETag('etag', False), None, 'W/"etag"', 412),
+        (ETag('etag', False), None, '"other", W/"etag"', 412),
+        (ETag('etag', False), None, '*', 412),
+        (ETag('etag', False), None, '"other"', 200),
+
+        (ETag('etag'), '"etag"', '"etag"', 412),
+        (ETag('etag'), '"etag"', 'W/"etag"', 412),
+        (ETag('etag'), '"etag"', '"other"', 200),
+        (ETag('etag'), 'W/"etag"', '"etag"', 412),
+        (ETag('etag'), 'W/"etag"', '"other"', 412),
+    ]
+)
+def test_process_conditional_put_requests(web_app, pyramid_request, app_config,
+                                          etag, if_match, if_none_match, status_code):
+    app_config.scan('restfw.tests.test_viewderivers')
+    app_config.commit()
+    root = pyramid_request.root
+    root['resource'] = resource = DummyResource({'foo': 'Hello', 'bar': 123})
+    resource.etag = etag
+    resource_url = pyramid_request.resource_url(root['resource'])
+
+    headers = {}
+    if if_match:
+        headers['If-Match'] = if_match
+    if if_none_match:
+        headers['If-None-Match'] = if_none_match
+
+    kwargs = {'headers': headers}
+    if status_code == 304:
+        kwargs['exception'] = HTTPNotModified()
+    elif status_code == 412:
+        kwargs['exception'] = HTTPPreconditionFailed({'etag': None if etag is None else etag.serialize()})
+    else:
+        kwargs['status'] = status_code
+
+    new_params = {'foo': 'World', 'bar': 456}
+    web_app.put_json(resource_url, params=new_params, **kwargs)